+++ projects/dts-backoffice-kendo-grid/src/lib/components/dts-kendo-grid/dts-kendo-grid-column.interface.ts
@@ -14,65 +14,53 @@
 }
 
 export interface DtsKendoGridColumn {
-    /**
-     * Nome do atributo do JSON passado no atributo "t-data" do componente
-     * que representerá o valor a ser exibido na coluna.
-     */
+    /** Nome do atributo do JSON passado no atributo "d-data" do componente. */
     column: string;
 
     /** Nome da coluna a ser exibido na tabela. */
     label: string;
 
-    /** Tamanho da coluna em pixels. */
-    width?: number;
-
-
     /**
-     * Informa se a coluna está agrupada.
-     *
-     * Ao iniciar o Totvs Grid com essa propriedade setada como `true`, a coluna iniciará agrupada automaticamente,
-     * por padrão todas as colunas vem desagrupadas (`false`).
-     *
-     * Importante: Para agrupar uma coluna é necessário que a propriedade `t-groupable` do componente Totvs Grid esteja setada como
-     * `true`, caso contrário, esta configuração será ignorada.
-     */
-    groupHeader?: boolean;
-
-    /** Se for passado "true" é habilitado o componente checkbox na coluna. */
-    checkbox?: boolean;
-
-    /** Se for passado o valor "true" é habilitado o campo de edição no modo edição habilitado. */
-    editable?: boolean;
-
-    /** Se for passado o valor "true" o campo se torna obrigatório no modo edição habilitado. */
-    required?: boolean;
-
-    /**
-     * Define o tipo de campo em modo de edição. Valores válidos:
-     *
-     * - **`string` (padrão)**: campo do tipo texto;
-     * - **`number`**: campo do tipo numérico;
-     * - **`date`**: campo do tipo date;
-     * - **`currency`**: campo do tipo monetário.
-     * - **`label`**: campo do tipo label.
+     * Define o tipo da coluna. Os Tipos válidos são:
+     * - **`string` (padrão)**: Campo do tipo texto.
+     * - **`number`**: Campo do tipo numérico.
+     * - **`currency`**: Campo do tipo monetário.
+     * - **`boolean`**: Campo do tipo lógico.
+     * - **`date`**: Campo do tipo data.
+     * - **`label`**: Campo do tipo label (apresenta uma TAG).
+     * - **`subtitle`**: Campo do tipo label (apresenta uma lista de TAG's).
      */
     type?: string;
 
-    /**
-     * Formata os dados da coluna de acordo com o tipo. Segue abaixo exemplos de preenchimento:
-     *
-     * - **Formato para data (`date`)**: aceita apenas os caracteres de dia(dd), mês(MM ou mm) e ano (yyyy ou yy),
-     * valor padrão é 'dd/MM/yyyy'. Exemplos: 'dd/MM/yyyy', 'dd-MM-yy', 'mm/dd/yyyy'.
-     */
+    /** Formata os dados da coluna de acordo com o tipo. */
     format?: string;
 
-    labels?: Array<DtsLabel>;
+    /** Tamanho da coluna em pixels. */
+    width?: number;
 
-    // Para formatar moeda com o PIPE do angular
-    // https://angular.io/api/common/CurrencyPipe
+    /** Se definido e igual a "true", campos do tipo 'boolean' são apresentado como checkbox, senão, apresenta como texto (Sim/Não). */
+    checkbox?: boolean;
+
+    /** Moeda utilizada para formatar campos do tipo 'currency' usando o Currency PIPE (https://angular.io/api/common/CurrencyPipe). */
     currency?: string;
 
+    /** Formato utilizada para formatar campos do tipo 'currency' usando o Currency PIPE (https://angular.io/api/common/CurrencyPipe). */
     symbol?: string;
+
+    /** Lista de valores que devem ser exibidos quando o tipo é label ou subtitle. O objeto da lista utiliza a interface DtsLabel. */
+    labels?: Array<DtsLabel>;
+
+    /** Indica se o campo deve ser hablitado para edição. O parâmetro "d-editable" do Grid deve ser igual a "true". */
+    editable?: boolean;
+
+    /** Indica se o campo é obrigatório na edição. Os parâmetros "editable" (coluna) o "d-editable" (Grid) devem ser igual a "true". */
+    required?: boolean;
+
+    /** Indica se a coluna deve ser agrupada na inicialização. O parâmetro "d-groupable" do Grid deve ser igual a "true". */
+    groupHeader?: boolean;
+
+    /** Indica se a coluna está visível ou não */
+    visible?: boolean;
 }
 
 export interface DtsEditAction {